Implementing Login Features per Level

+++ app/Http/Controllers/BarangController.php
@@ -8,6 +8,7 @@ use App\Models\Barang;
 use App\Models\JenisPengadaan;
 use App\Models\Kategorial;
 use App\Models\Ruangan;
+use App\Models\Unit;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
 
@@ -27,14 +28,18 @@ class BarangController extends Controller
 
     public function create()
     {
+        // $this->authorize('create', Barang::class);
         $ruangans = Ruangan::all();
         $jenis_pengadaans = JenisPengadaan::all();
-        $kategorials = Kategorial::all(); // Tambahkan ini untuk mendapatkan data kategorial
-        return view('barangs.create', compact('ruangans', 'jenis_pengadaans', 'kategorials'));
+        $kategorials = Kategorial::all();
+        $units = Unit::all();
+        return view('barangs.create', compact('ruangans', 'jenis_pengadaans', 'kategorials', 'units'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Barang::class);
+
         if ($request->hasFile('gambar_barang')) {
             $path = $request->file('gambar_barang')->store('uploads');
         }
@@ -45,6 +50,9 @@ class BarangController extends Controller
             'tipe' => $request->tipe,
             'catatan' => $request->catatan,
             'ruangan_id' => $request->ruangan_id,
+            'kategorial_id' => $request->kategori_id,
+            'jenis_pengadaan_id' => $request->jenis_pengadaan_id,
+            'unit_id' => $request->unit_id,
             'tahun' => $request->tahun,
             'kondisi' => $request->kondisi,
             'jumlah' => $request->jumlah,
@@ -66,31 +74,22 @@ class BarangController extends Controller
         $barang = Barang::find($id);
         $ruangans = Ruangan::all();
         $jenis_pengadaans = JenisPengadaan::all();
-        $kategorials = Kategorial::all(); // Tambahkan ini untuk mendapatkan data kategorial
-        return view('barangs.edit', compact('barang', 'ruangans', 'jenis_pengadaans', 'kategorials'));
+        $kategorials = Kategorial::all();
+        $units = Unit::all();
+        return view('barangs.edit', compact('barang', 'ruangans', 'jenis_pengadaans', 'kategorials', 'units'));
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, Barang $barang)
     {
-        $barang = Barang::findOrFail($id);
-
-        // Simpan jumlah barang sebelum diupdate
-        $jumlahSebelum = $barang->jumlah;
-
+        $this->authorize('update', Barang::class);
+        // $barang = Barang::findOrFail($id);
         $barang->update($request->all());
-
-        // Hitung selisih jumlah barang sebelum dan setelah diupdate
-        $selisihJumlah = $jumlahSebelum - $request->jumlah;
-
-        if ($selisihJumlah != 0) {
-            $barang->jumlah -= $selisihJumlah;
-            $barang->save();
-        }
         return redirect()->route('barangs.index')->with('success', 'Barang berhasil diperbarui.');
     }
 
     public function destroy($id)
     {
+        $this->authorize('destroy', Barang::class);
         $barang = Barang::find($id);
         $barang->delete();
         return redirect()->route('barangs.index')->with('success', 'Barang berhasil dihapus.');
